+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 * Updated detailed description for community integrations.
 * Added support for special fields: *defaultclassifier*, *defaultmapperin*, *defaultmapperout* in **download** command.
+* Added -y option **format** command to assume "yes" as answer to all prompts and run non-interactively
 
 # 1.2.3
 * Fixed an issue where running **validate** failed on playbook with task that adds tags to the evidence data.